// frontend/src/app/types/expanse.ts
export interface NewExpanseType {
    title: string;
    description: string;
    category: string;
    price: number;
    date: string;
    id?: string;
  }

  export interface NewIncomeType {
    title: string;
    description: string;
    category: string;
    price: number;
    date: string;
    id?: string;
  }
  
  export interface AddExpanseType {
    title: string;
    description: string;
    category: string;
    price: number;
    date: string;
  }

  export interface AddIncomeType {
    title: string;
    description: string;
    category: string;
    price: number;
    date: string;
  }
  